static files code fixed and moved
